Features:   - Create separate publications app (renamed from blog)   - Add static Django index pages at /blog/ and /events/ (no CMS setup needed)   - Implement prefixed URLs: /blog/slug/ and /events/slug/   - Separate BlogPage and EventPage models completely   - Add individual page templates for both content types
