+++ core/urls.py
@@ -5,4 +5,8 @@ app_name = 'core'
 
 urlpatterns = [
     path('', views.home, name='home'),
+    path('blog/', views.blog_index, name='blog_index'),
+    path('events/', views.events_index, name='events_index'),
+    path('blog/<slug:slug>/', views.serve_blog_page, name='blog_page'),
+    path('events/<slug:slug>/', views.serve_event_page, name='event_page'),
 ]
