+++ core/views.py
@@ -1,12 +1,13 @@
-from django.shortcuts import render
+from django.shortcuts import render, get_object_or_404
 import logging
 
 from .utils import (
-    get_recent_contributors, 
+    get_recent_contributors,
     get_recent_packages,
-    generate_github_avatar_url, 
+    generate_github_avatar_url,
     generate_github_profile_url
 )
+from publications.models import BlogPage, EventPage
 
 logger = logging.getLogger(__name__)
 
@@ -59,3 +60,99 @@ def home(request):
         'recent_packages': recent_packages,
     }
     return render(request, 'core/home.html', context)
+
+
+def blog_index(request):
+    """
+    Blog index view for PyOpenSci.
+
+    Static Django view that queries Wagtail BlogPage instances
+    to display all blog posts in a consistent index page.
+
+    Parameters
+    ----------
+    request : HttpRequest
+        Django HTTP request object.
+
+    Returns
+    -------
+    HttpResponse
+        Rendered blog index page with blog posts.
+    """
+    blog_posts = BlogPage.objects.live().select_related('author').order_by('-date')
+
+    context = {
+        'page_title': 'pyOpenSci Blog',
+        'hero_title': 'pyOpenSci Blog',
+        'hero_subtitle': 'Here we will both post updates about pyOpenSci and also highlight contributors. We will also highlight new packages that have been reviewed and accepted into the pyOpenSci ecosystem.',
+        'blog_posts': blog_posts,
+    }
+    return render(request, 'core/blog_index.html', context)
+
+
+def events_index(request):
+    """
+    Events index view for PyOpenSci.
+
+    Static Django view that queries Wagtail EventPage instances
+    to display all events in a consistent index page.
+
+    Parameters
+    ----------
+    request : HttpRequest
+        Django HTTP request object.
+
+    Returns
+    -------
+    HttpResponse
+        Rendered events index page with events.
+    """
+    events = EventPage.objects.live().select_related('author').prefetch_related('tags').order_by('-start_date')
+
+    context = {
+        'page_title': 'pyOpenSci Events',
+        'hero_title': 'pyOpenSci Events',
+        'hero_subtitle': 'Join us for workshops, webinars, and community events. Connect with the scientific Python community and learn about open source best practices.',
+        'events': events,
+    }
+    return render(request, 'core/events_index.html', context)
+
+
+def serve_blog_page(request, slug):
+    """
+    Serve individual blog page with /blog/ prefix.
+
+    Parameters
+    ----------
+    request : HttpRequest
+        Django HTTP request object
+    slug : str
+        Page slug from URL
+
+    Returns
+    -------
+    HttpResponse
+        Rendered blog page using Wagtail's serve mechanism
+    """
+    page = get_object_or_404(BlogPage.objects.live().select_related('author'), slug=slug)
+    return page.serve(request)
+
+
+def serve_event_page(request, slug):
+    """
+    Serve individual event page with /events/ prefix.
+
+    Parameters
+    ----------
+    request : HttpRequest
+        Django HTTP request object
+    slug : str
+        Page slug from URL
+
+    Returns
+    -------
+    HttpResponse
+        Rendered event page using Wagtail's serve mechanism
+    """
+    page = get_object_or_404(EventPage.objects.live().select_related('author').prefetch_related('tags'), slug=slug)
+    return page.serve(request)
