+++ templates/base.html
@@ -70,7 +70,7 @@
                             <div class="py-2">
                                 <a href="#" class="block px-4 py-2 text-sm text-gray-700 hover:bg-pyos-light-purple hover:text-pyos-deep-purple">Tutorials & Resources</a>
                                 <a href="/blog/" class="block px-4 py-2 text-sm text-gray-700 hover:bg-pyos-light-purple hover:text-pyos-deep-purple">Blog</a>
-                                <a href="#" class="block px-4 py-2 text-sm text-gray-700 hover:bg-pyos-light-purple hover:text-pyos-deep-purple">Events</a>
+                                <a href="/events/" class="block px-4 py-2 text-sm text-gray-700 hover:bg-pyos-light-purple hover:text-pyos-deep-purple">Events</a>
                                 <a href="#" class="flex items-center px-4 py-2 text-sm text-gray-700 hover:bg-pyos-light-purple hover:text-pyos-deep-purple">Ask a question <i class="fa-brands fa-discourse ml-2 text-xs"></i></a>
                             </div>
                         </div>
